+++ tests/bhyveargv2xmltest.c
@@ -76,7 +76,7 @@ testCompareXMLToArgvFiles(const char *xmlfile,
         return -1;
     }
 
-    if (vmdef && !(actualxml = virDomainDefFormat(vmdef, driver.xmlopt, 0)))
+    if (vmdef && !(actualxml = virDomainDefFormat(vmdef, driver.xmlopt, VIR_DOMAIN_DEF_FORMAT_SECURE)))
         return -1;
 
     if (vmdef && virTestCompareToFile(actualxml, xmlfile) < 0)
@@ -187,6 +187,7 @@ mymain(void)
     DO_TEST("vnc-vga-off");
     DO_TEST("vnc-vga-io");
     DO_TEST("vnc-resolution");
+    DO_TEST("vnc-password");
 
     virObjectUnref(driver.caps);
     virObjectUnref(driver.xmlopt);